# typescript-sdk/openapi/radioso.yaml
openapi: 3.1.0
info:
  title: radioso API
  version: 0.1.0
  description: Code-generated OpenAPI contract for the radioso backend
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: radioso_session
    workspaceSelection:
      type: apiKey
      in: header
      name: X-Workspace-Id
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: APIKey
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
          required:
            - code
            - message
      required:
        - error
    FlatErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
    UpdateRetrievalSettingsRequest:
      type: object
      properties:
        queryRewriteEnabled:
          type: boolean
        semanticRewriteInstructions:
          type: string
          maxLength: 2000
        lexicalRewriteInstructions:
          type: string
          maxLength: 2000
        suggestedQuestionsEnabled:
          type: boolean
        rerankEnabled:
          type: boolean
        vectorTopK:
          type: integer
          minimum: 1
        similarityThreshold:
          type: number
        rerankTopK:
          type: integer
          maximum: 50
        citationDisplayEnabled:
          type: boolean
        metadataRules:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                minLength: 1
              field:
                type: string
                minLength: 1
              valueType:
                type: string
                enum: *a1
              operator:
                type: string
                enum: *a2
              value:
                type: string
              combinator:
                type: string
                enum: &a5
                  - and
                  - or
              conditions:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      minLength: 1
                    field:
                      type: string
                      minLength: 1
                    valueType:
                      type: string
                      enum: *a1
                    operator:
                      type: string
                      enum: *a2
                    value:
                      type: string
                  required:
                    - id
                    - field
                    - valueType
                    - operator
                    - value
              effect:
                type: string
                enum: *a3
              enabled:
                type: boolean
              triggerMode:
                type: string
                enum: *a4
              triggerInstruction:
                type: string
                maxLength: 500
            required:
              - id
              - effect
              - enabled
        customInstruction:
          type: string
          maxLength: 2000
      required:
        - queryRewriteEnabled
        - rerankEnabled
        - vectorTopK
        - similarityThreshold
        - rerankTopK
        - citationDisplayEnabled
...
    PlatformSettingsResponse:
      type: object
      properties:
        assistant:
          $ref: "#/components/schemas/AssistantSettingsSection"
        retrieval:
          $ref: "#/components/schemas/PlatformRetrievalSettingsSection"
        channels:
          $ref: "#/components/schemas/PlatformChannelsSettingsSection"
      required:
        - assistant
        - retrieval
        - channels
...
    WorkspaceIngestionReprocessResponse:
      type: object
      properties:
        workspaceId:
          type: string
          format: uuid
        queuedDocumentCount:
          type: integer
          minimum: 0
        skippedDocumentCount:
          type: integer
          minimum: 0
        status:
          type: string
          enum:
            - queued
            - noop
      required:
        - workspaceId
        - queuedDocumentCount
        - skippedDocumentCount
        - status
...
paths:
...
  /api/v1/settings:
    get:
      tags:
        - Settings
      summary: Get shared workspace platform settings
      operationId: getPlatformSettings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Shared assistant, retrieval, and channel settings returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettingsResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - Settings
      summary: Merge-update shared workspace platform settings
      operationId: updatePlatformSettings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlatformSettingsRequest"
      responses:
        "200":
          description: Shared settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettingsResponse"
        "400":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/settings/retrieval:
    get:
      tags:
        - Settings
      summary: Get retrieval settings for the authenticated workspace
      operationId: getRetrievalSettings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Retrieval settings returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetrievalSettings"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - Settings
      summary: Update retrieval settings for the authenticated workspace
      operationId: updateRetrievalSettings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRetrievalSettingsRequest"
      responses:
        "200":
          description: Updated retrieval settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetrievalSettings"
        "400":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/settings/ingestion:
    get:
      tags:
        - Settings
      summary: Get ingestion settings for the authenticated workspace
      operationId: getIngestionSettings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Ingestion settings returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionSettings"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - Settings
      summary: Update ingestion settings for the authenticated workspace
      operationId: updateIngestionSettings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIngestionSettingsRequest"
      responses:
        "200":
          description: Updated ingestion settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionSettings"
        "400":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/settings/ingestion/embedding-model/cancel:
    post:
      tags:
        - Settings
      summary: Cancel a pending workspace embedding model change
      operationId: cancelPendingEmbeddingModel
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pending embedding model change cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionSettings"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/settings/ingestion/reprocess:
    post:
      tags:
        - Settings
      summary: Queue eligible workspace documents for reprocessing using current ingestion settings
      operationId: reprocessWorkspaceIngestion
      security:
        - bearerAuth: []
      responses:
        "202":
          description: Workspace documents accepted for reprocessing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceIngestionReprocessResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/settings/general:
    get:
      tags:
        - Settings
      summary: Get general workspace settings
      operationId: getGeneralSettings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: General settings returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralSettingsResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlatErrorResponse"
    put:
      tags:
        - Settings
      summary: Update general workspace settings
      operationId: updateGeneralSettings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGeneralSettingsRequest"
      responses:
        "200":
          description: Updated general settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralSettingsResponse"
        "400":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlatErrorResponse"
  /api/v1/settings/general/anonymous-chat-token/rotate:
    post:
      tags:
        - Settings
      summary: Reset the anonymous chat public link
      operationId: rotateAnonymousChatToken
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Updated general settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralSettingsResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/settings/general/website-embed-token/rotate:
    post:
      tags:
        - Settings
      summary: Reset the website embed token
      operationId: rotateWebsiteEmbedToken
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Updated general settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralSettingsResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/settings/general/assistant-logo:
    post:
      tags:
        - Settings
      summary: Upload the default assistant logo
      operationId: uploadAssistantLogo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/AssistantLogoUploadRequest"
      responses:
        "200":
          description: Updated general settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralSettingsResponse"
        "400":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Settings
      summary: Remove the default assistant logo
      operationId: deleteAssistantLogo
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Updated general settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralSettingsResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...